refactor: split github.ts service into focused modules (GMC-28)
Break the ~1,870-line github.ts monolith into cohesive modules behind a
barrel re-export, with no behaviour change. All existing imports from
`services/github` resolve unchanged via `github/index.ts`.

Modules:
- client     — Octokit + persisted Apollo client setup (shared singleton)
- queries    — read queries + RepoNode normalisation
- mutations  — repo write operations (star, archive, visibility, rename)
- rest       — REST-only operations (delete, create, transfer, sync, rate limits)
- cache      — Apollo cache reads/updates/eviction + cache file inspection
- enrichment — batched fork ahead/behind enrichment

`makeApolloClient` and its singleton live solely in `client` so the cached
Apollo instance stays a true singleton across all modules.

Typecheck clean; full suite 376/376.
